+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Change rake task namespace from `semantic_release` to `release` and make configurable
+
 ## 1.0.1 (02 February 2025)
 
 - Add blank line after version number in changelog